user managed by super class
